server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.4.2
opencv-python-headless>=4.8.0
mediapipe-silicon==0.10.8